refactor(AddInformationFields): separated logic via custom hooks
