+++ src/components/AddInformationFields/useComponentRules.ts
@@ -1,0 +1,33 @@
+import { useCallback } from 'react'
+import { Control, FieldValues } from 'react-hook-form'
+
+interface UseComponentRulesProps {
+  name: string
+  list: string[]
+  control: Control<FieldValues>
+  setList(list: string[]): void
+}
+
+const useComponentRules = ({ name, list, control, setList }: UseComponentRulesProps) => {
+  const handleAdd = useCallback(() => setList([...list, '']), [list])
+
+  const handleRemove = useCallback(
+    (index: number) => {
+      const formData = control.getValues()
+      const listFiltered: string[] = formData[name].filter((_: any, _index: number) => _index !== index)
+
+      if (!listFiltered.length) {
+        listFiltered.push('')
+      }
+
+      listFiltered.forEach((value, i) => control.setValue(`${name}[${i}]`, value))
+
+      setList(listFiltered)
+    },
+    [list, name]
+  )
+
+  return { handleAdd, handleRemove }
+}
+
+export default useComponentRules
